requirements.txt
flask
requests
APScheduler
pyYaml
colorama